Changed createdAt to a BigInt
