+++ src/core/announcements/crypto.test.ts
@@ -15,7 +15,7 @@ describe("announcement crypto", () => {
     });
 
     it("returns a valid signature for a valid private key and announcement", async () => {
-      const announcement = createBroadcast("1", "https://dsnp.org", "0x12345", 1627324121352);
+      const announcement = createBroadcast("1", "https://dsnp.org", "0x12345", BigInt(1627324121352));
       const signedAnnouncement = await sign(announcement);
 
       expect(signedAnnouncement.signature).toEqual(
@@ -28,7 +28,7 @@ describe("announcement crypto", () => {
     const publicKey = "0x307FBB8320C719B7A4Cd1F9215d8927C28066583";
 
     it("returns the correct public key for a valid signature", () => {
-      const announcement = createBroadcast("1", "https://dsnp.org", "0x12345", 1627324121352);
+      const announcement = createBroadcast("1", "https://dsnp.org", "0x12345", BigInt(1627324121352));
       const signature =
         "0xd33f14693809e6c7bcd5148cb585a63ce51d54bd229a7306dab22d3437b001140538494c1b7b19a2a806bcc6da26fc205b237cfe4a60dd738d994ec72e2a6a561c";
 
@@ -36,7 +36,7 @@ describe("announcement crypto", () => {
     });
 
     it("returns a different public key for an invalid signature", () => {
-      const announcement = createBroadcast("1", "https://dsnp.org", "0x12345", 1627324121352);
+      const announcement = createBroadcast("1", "https://dsnp.org", "0x12345", BigInt(1627324121352));
       const signature =
         "0x168aeea67ab8e6b60a8d004ab57e17bab16dac135bcc9d9c8b34058808befd371e5d34ee6b1dc18b3d86e63765352286ba949a39cd458138cad7895b7faca2cb1c";
 
@@ -65,7 +65,7 @@ describe("announcement crypto", () => {
       fromId: "0x12345",
       contentHash: "0x67890",
       url: "https://www.dsnp.org/",
-      createdAt: +new Date("2021-07-31T10:11:12Z"),
+      createdAt: BigInt(+new Date("2021-07-31T10:11:12Z")),
     };
 
     const privateKey = "0xd9d3b5afb7765ffd9f047fd0d1d9b47d4d538b6a56f1cf29dc160ab9c6d30aa3";
