+++ src/createPublication.test.ts
@@ -13,7 +13,7 @@ describe("createPublication", () => {
       "dsnp://0x0123456789ABCDEF",
       "https://dsnp.org",
       "0x0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF",
-      1627324121352
+      BigInt(1627324121352)
     ),
   ];
 
@@ -67,32 +67,32 @@ describe("createPublications", () => {
       "dsnp://0x0123456789ABCDEF",
       "https://dsnp.org",
       "0x0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF",
-      1627324121352
+      BigInt(1627324121352)
     ),
     createBroadcast(
       "dsnp://0x0123456789ABCDE0",
       "https://dsnp.org",
       "0x1123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF",
-      1627324121352
+      BigInt(1627324121352)
     ),
     createReply(
       "dsnp://0x0123456789ABCDEF",
       "https://dsnp.org",
       "0x0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF",
       "dsnp://0x0123456789ABCDEF/0x0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF",
-      1627324121352
+      BigInt(1627324121352)
     ),
     createReaction(
       "dsnp://0x0123456789ABCDEF",
       "🏳️‍🌈",
       "dsnp://0x0123456789ABCDEF/0x0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF",
-      1627324121352
+      BigInt(1627324121352)
     ),
     createReaction(
       "dsnp://0x0123456789ABCDEF",
       "🏳️‍🌈",
       "dsnp://0x0123456789ABCDE0/0x0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF",
-      1627324121352
+      BigInt(1627324121352)
     ),
   ];
